backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
aiofiles==0.7.0
    # via piccolo-admin
aiosqlite==0.17.0
    # via piccolo
anyio==3.3.4
    # via
    #   httpcore
    #   starlette
argcomplete==1.12.3
    # via datamodel-code-generator
asgiref==3.4.1
    # via uvicorn
asyncpg==0.25.0
    # via piccolo
attrs==21.2.0
    # via jsonschema
black==21.11b1
    # via
    #   -r requirements-dev.in
    #   datamodel-code-generator
    #   piccolo
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==4.0.0
    # via prance
charset-normalizer==2.0.7
    # via
    #   httpx
    #   requests
click==7.1.2
    # via
    #   black
    #   typer
    #   uvicorn
colorama==0.4.4
    # via
    #   piccolo
    #   targ
    #   typer
datamodel-code-generator[http]==0.11.9
    # via fastapi-code-generator
dnspython==2.1.0
    # via email-validator
docstring-parser==0.12
    # via targ
email-validator==1.1.3
    # via pydantic
fastapi==0.70.0
    # via
    #   -r requirements.in
    #   piccolo-admin
    #   piccolo-api
fastapi-code-generator==0.3.0
    # via -r requirements.in
genson==1.2.2
    # via datamodel-code-generator
h11==0.12.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==0.14.3
    # via httpx
httptools==0.2.0
    # via uvicorn
httpx==0.21.1
    # via datamodel-code-generator
hypercorn==0.12.0
    # via piccolo-admin
hyperframe==6.0.1
    # via h2
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
inflect==5.3.0
    # via datamodel-code-generator
inflection==0.5.1
    # via piccolo
isodate==0.6.0
    # via openapi-schema-validator
isort==5.10.1
    # via
    #   -r requirements-dev.in
    #   datamodel-code-generator
jinja2==3.0.3
    # via
    #   datamodel-code-generator
    #   fastapi-code-generator
    #   piccolo
    #   piccolo-api
jsonschema==4.2.1
    # via
    #   openapi-schema-validator
    #   openapi-spec-validator
markupsafe==2.0.1
    # via jinja2
mypy-extensions==0.4.3
    # via black
openapi-schema-validator==0.1.5
    # via openapi-spec-validator
openapi-spec-validator==0.3.1
    # via datamodel-code-generator
pathspec==0.9.0
    # via black
piccolo[postgres,sqlite]==0.60.0
    # via
    #   -r requirements.in
    #   piccolo-admin
    #   piccolo-api
piccolo-admin==0.18.1
    # via -r requirements.in
piccolo-api==0.29.0
    # via
    #   -r requirements.in
    #   piccolo-admin
platformdirs==2.4.0
    # via black
prance==0.21.8.0
    # via datamodel-code-generator
priority==2.0.0
    # via hypercorn
pydantic[email]==1.8.2
    # via
    #   datamodel-code-generator
    #   fastapi
    #   fastapi-code-generator
    #   piccolo
    #   piccolo-api
pyjwt==2.3.0
    # via piccolo-api
pyrsistent==0.18.0
    # via jsonschema
pysnooper==0.5.0
    # via
    #   datamodel-code-generator
    #   fastapi-code-generator
python-dotenv==0.19.2
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.5
    # via piccolo-api
pyyaml==6.0
    # via
    #   openapi-spec-validator
    #   uvicorn
regex==2021.11.10
    # via black
requests==2.26.0
    # via prance
rfc3986[idna2008]==1.5.0
    # via httpx
ruamel.yaml==0.17.17
    # via prance
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
semver==2.13.0
    # via prance
shellingham==1.4.0
    # via typer
six==1.16.0
    # via
    #   openapi-schema-validator
    #   openapi-spec-validator
    #   prance
    #   python-multipart
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.16.0
    # via fastapi
stringcase==1.2.0
    # via fastapi-code-generator
targ==0.3.6
    # via
    #   piccolo
    #   piccolo-admin
toml==0.10.2
    # via
    #   datamodel-code-generator
    #   hypercorn
tomli==1.2.2
    # via black
typer[all]==0.3.2
    # via fastapi-code-generator
typing-extensions==4.0.0
    # via
    #   aiosqlite
    #   black
    #   piccolo
    #   pydantic
urllib3==1.26.7
    # via requests
uvicorn[standard]==0.15.0
    # via
    #   -r requirements.in
    #   piccolo-admin
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn
wsproto==1.0.0
    # via hypercorn
